Splitting tests.py on 2 test files, adding conftest for fixtures
